refactor(api): move filter util
