+++ api/src/prescription/campaign/infrastructure/repositories/campaign-report-repository.js
@@ -4,8 +4,8 @@ import { knex } from '../../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { CampaignParticipationStatuses } from '../../../../../lib/domain/models/index.js';
 import { CampaignReport } from '../../../../../lib/domain/read-models/CampaignReport.js';
-import { filterByFullName } from '../../../../../lib/infrastructure/utils/filter-utils.js';
 import * as skillRepository from '../../../../shared/infrastructure/repositories/skill-repository.js';
+import { filterByFullName } from '../../../../shared/infrastructure/utils/filter-utils.js';
 import { fetchPage } from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { TargetProfileForSpecifier } from '../../../target-profile/domain/read-models/TargetProfileForSpecifier.js';
 import { getLatestParticipationSharedForOneLearner } from './helpers/get-latest-participation-shared-for-one-learner.js';
